Moves the controller actions in a own lib class.
